+++ dto/userinfo_response.go
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2019 Authlete, Inc.
+// Copyright (C) 2019-2020 Authlete, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -48,4 +48,7 @@ type UserInfoResponse struct {
 
 	//
 	ClientIdAliasUsed bool `json:"clientIdAliasUsed,omitempty"`
+
+	//
+	UserInfoClaims string `json:"userInfoClaims,omitempty"`
 }